requirements.txt
streamlit
llama-index
llama-index-embeddings-huggingface
llama-index-llms-langchain
python-dotenv
duckduckgo-search
llama-index-core
pandas
numpy
langchain
langchain-experimental
langchain-groq